Remove distinction between userAction{StartToolAt:,MoveToolTo:,StopTool} and {startToolAt:,moveToolTo:,stopTool}
This can be done because the new user action locking system prevents user actions from being created while they are restored during undo,
which was the (hopefully sole) original reason for the distinction.

This also removes the addition of user actions for these methods, which is the responsibility of the tools from now on.